+++ sqlalchemy_unchained/utils.py
@@ -1,6 +1,9 @@
 import re
+import sqlalchemy as sa
 
 from functools import reduce
+from py_meta_utils import McsArgs
+from typing import *
 
 
 _missing = type('_missing', (), {'__bool__': lambda self: False})()
@@ -82,6 +85,26 @@ def rec_hasattr(obj, attr):
         return False
     else:
         return True
+
+
+def _add_arg_to_table_args(mcs_args: McsArgs, new_arg: Any) -> Tuple:
+    table_args = mcs_args.clsdict.get('__table_args__', ())
+    if isinstance(table_args, dict):
+        table_args = (table_args,)
+
+    if table_args and isinstance(table_args[-1], dict):
+        new_table_args = *table_args[:-1], new_arg, table_args[-1]
+    else:
+        new_table_args = *table_args, new_arg
+    mcs_args.clsdict['__table_args__'] = new_table_args
+
+    return new_table_args
+
+
+def _get_column_names(mcs_args: McsArgs) -> Set[str]:
+    return {col.name or attr_name
+            for attr_name, col in mcs_args.clsdict.items()
+            if isinstance(col, sa.Column)}
 
 
 __all__ = [